Refactored a bit the compiler region so that it will be easier for bundler plugins to pick up the common functionalities (eg. linking)
