+++ test/general/specs/SolidityCompiler.node.spec.ts
@@ -4,7 +4,7 @@ describe("SolidityCompiler.node", () => {
   it("loading the solc compiler should not register the 'unhandledRejection' process event", async () => {
     const preLoadEventHandlersCount =
       process.listenerCount("unhandledRejection");
-    await import("../../../lib/SolidityCompiler");
+    await import("../../../lib/compiler/SolidityCompiler");
     const postLoadEventHandlersCount =
       process.listenerCount("unhandledRejection");
